+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.78.0 (Unreleased)
 BUG FIXES:
 * cdn: fixed wrong documentation example
+* postgresql: fix `login` and `conn_limit` wrong behaviour in `yandex_mdb_postgresql_user`
 
 ENHANCEMENTS:
 vpc: allow usage of `yandex_vpc_gateway` in `yandex_vpc_route_table.static_route` as `gateway_id` next hop